rework: 42CTF PAM service
This commit fixes an url injection.
Until now, the /accounts/me route took a token as a parameter in the
form of /accounts/me/<token>.
The previous PAM implementation simply put the user's input inside the
url leading to a possible path traversal.

Thus, with an input such as '../../', the url would be
/accounts/me/../../ and the request succeed.