update: use supervisor to manage dockerd and sshd
Removing crappy management of sshd and dockerd.
The manager now spawn a process to run supervisor,
and will simply wait for SIGHUP to update sshd and dockerd.
The manager also use xmlrpc to communicate with supervisord and retrieve
infos about processes: if they have died etc.
If a process died for some reasons and isn't able to restart, then the
container exists.