update: lets the manager control docker images
Up to now, when a container was spawned, if the images didn't exist
locally, the spawner tried to pull the image and / or build the image
if a build context were specified.

The build informations were displayed to the user, leading to potential
leaks.
To avoid that, the manager can now build docker images on startup, and
on SIGHUP.