fix: error message in spawner entrypoint when images not ready
A message is displayed to the user if he tries to connect to a container
that has not yet been built.
Up to now, this message was displayed if docker run returns exit code.
But in fact, the docker run command will also return the entrypoint's
exit code.
Thus, we can't use ``docker run [...] || echo "Not available"``